% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities_list.R
\name{activities_list}
\alias{activities_list}
\title{Retrieve list of event log activities}
\usage{
activities_list(log)
}
\arguments{
\item{log}{An object of class log}
}
\description{
(maybe not export, for internal use only)
}
