% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outcome_labels.R
\name{assign_outcome_labels}
\alias{assign_outcome_labels}
\title{Assign column outcome with outcome labels based on last activity from the prefixed dataframe}
\usage{
assign_outcome_labels(prefix_df, ...)
}
\description{
(WIP)
}
