% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_fit_model.R
\name{compile_fit_model}
\alias{compile_fit_model}
\title{Compile and fit transformer model}
\usage{
compile_fit_model(
  transformer_model,
  tokens_train,
  learning_rate,
  num_epochs,
  batch_size,
  file
)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{tokens_train}{A list of train tokens, i.e. token_x, token_y}

\item{learning_rate}{A learning rate of a model}

\item{num_epochs}{A number of epochs}

\item{batch_size}{A batch size}

\item{file}{Name of saved model (weights)}
}
\description{
(WIP)
}
