% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{Compile and fit transformer model (test)}
\usage{
compile_model(transformer_model, learning_rate)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{learning_rate}{A learning rate of a model}
}
\description{
(test)
}
