% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model_pyversion.R
\name{compile_model_pyversion}
\alias{compile_model_pyversion}
\title{Compile transformer model}
\usage{
compile_model_pyversion(transformer_model, learning_rate)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{learning_rate}{A learning rate of a model}
}
\description{
Compile transformer model
}
