% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ProcessTransformer.R
\name{compile_ProcessTransformer}
\alias{compile_ProcessTransformer}
\title{Default compile function for ProcessTransformer model}
\usage{
compile_ProcessTransformer(object, ...)
}
\arguments{
\item{object}{Object of class \link{ppred_model}}

\item{...}{Additional Arguments}
}
\description{
Default compile function for ProcessTransformer model
}
