% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix for}
\usage{
confusion_matrix(y_pred, test_data, as.tibble = F)
}
\arguments{
\item{y_pred}{Predicted values}

\item{test_data}{Either a dataset or tokens}

\item{as.tibble}{\code{\link{logical}} (default \code{\link{FALSE}}): if \code{\link{TRUE}} returns a tibble}
}
\description{
Confusion matrix for
}
