% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vocabulary.R
\name{create_vocabulary}
\alias{create_vocabulary}
\title{Create a vocabulary, i.e. c("PAD", "UNK"), activities, outcome labels}
\usage{
create_vocabulary(processed_df)
}
\arguments{
\item{processed_df}{A processed dataframe}
}
\value{
a character vector of unique elements
}
\description{
Create a vocabulary, i.e. c("PAD", "UNK"), activities, outcome labels
}
