% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.ppred_model.R
\name{evaluate.ppred_model}
\alias{evaluate.ppred_model}
\title{Generic evaluate function for ppred_model class.}
\usage{
\method{evaluate}{ppred_model}(object, test_data, ...)
}
\arguments{
\item{object}{\code{\link{ppred_model}} (default \code{\link{NULL}}): ProcessTransformer model of class \code{\link{ppred_model}}.}

\item{test_data}{\code{\link{ppred_examples_df}} (default \code{\link{NULL}}): preprocessed test data.}

\item{...}{Additional arguments}
}
\description{
wip
}
