% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ProcessTransformer.R
\name{evaluate_ProcessTransformer}
\alias{evaluate_ProcessTransformer}
\title{Generic evaluate function for ppred_model class.}
\usage{
evaluate_ProcessTransformer(object, test_data, ...)
}
\arguments{
\item{object}{\code{\link{ppred_model}} (default \code{\link{NULL}}): ProcessTransformer model of class \code{\link{ppred_model}}.}

\item{test_data}{\code{\link{ppred_examples_df}} (default \code{\link{NULL}}): preprocessed test data.}

\item{...}{Additional arguments}
}
\description{
wip
}
