% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model (test)}
\usage{
fit_model(transformer_model, tokens_train, num_epochs, batch_size, file)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{tokens_train}{A list of train tokens, i.e. token_x, token_y}

\item{num_epochs}{A number of epochs}

\item{batch_size}{A batch size}

\item{file}{Name of saved model (weights)}
}
\description{
(test)
}
