% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_pyversion.R
\name{fit_model_pyversion}
\alias{fit_model_pyversion}
\title{Fit model (test)}
\usage{
fit_model_pyversion(
  transformer_model,
  train_data,
  num_epochs,
  batch_size,
  output_folder
)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{train_data}{A training dataset}

\item{num_epochs}{A number of epochs}

\item{batch_size}{A batch size}

\item{output_folder}{Name of saved model (weights)}
}
\description{
(test)
}
