% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ProcessTransformer.R
\name{fit_ProcessTransformer}
\alias{fit_ProcessTransformer}
\title{Default fit function for ProcessTransformer model}
\usage{
fit_ProcessTransformer(
  object,
  train_data,
  batch_size = batch_size,
  epochs = epochs,
  verbose = 1,
  callbacks = callbacks,
  shuffle = TRUE,
  validation_split = 0.2,
  ...
)
}
\arguments{
\item{object}{Object of class \link{ppred_model}: A transformer model}

\item{train_data}{A training dataset}

\item{batch_size}{A batch size}

\item{verbose}{A verbose}

\item{callbacks}{\code{\link{list}}: A list of callbacks}

\item{shuffle}{\code{\link{logical}} (default \code{\link{TRUE}}): If \code{\link{TRUE}} shuffles the data}

\item{validation_split}{A ratio to split on}

\item{...}{Additional Arguments}

\item{num_epochs}{A number of epochs}
}
\description{
Default fit function for ProcessTransformer model
}
