% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_case_length.R
\name{max_case_length}
\alias{max_case_length}
\title{Calculate the maximum length of a case / number of activities in the longest trace in an event log.}
\usage{
max_case_length(processed_df)
}
\arguments{
\item{processed_df}{A preprocessed dataframe from create_prefix_df().}
}
\value{
An integer number of the maximum case length (longest trace) in an event log.
}
\description{
(WIP)
}
