% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ProcessTransformer.R
\name{model_ProcessTransformer}
\alias{model_ProcessTransformer}
\title{Define transformer model}
\usage{
model_ProcessTransformer(df, ...)
}
\arguments{
\item{df}{\code{\link{data.frame}}: A processed \code{\link{data.frame}} from prepare_examples}
}
\description{
Defines the model using the keras functional API.
}
