% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_outputs.R
\name{num_outputs}
\alias{num_outputs}
\title{Calculate number of outputs (target variables)}
\usage{
num_outputs(processed_df)
}
\arguments{
\item{processed_df}{A processed dataset of class \code{\link{ppred_examples_df}}.}
}
\value{
an \code{integer} number of outputs for supplying as an argument to a Transformer model, i.e. number of unique labels for a specific process monitoring task.
}
\description{
Calculate number of outputs (target variables)
}
\examples{
library(processpredictR)
library(eventdataR)
df <- prepare_examples(patients)
num_outputs(df)

}
