% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ppred_predictions.R
\name{plot.ppred_predictions}
\alias{plot.ppred_predictions}
\title{Generic plot function}
\usage{
\method{plot}{ppred_predictions}(predictions, ...)
}
\arguments{
\item{predictions:}{An object of class \code{\link{ppred_predictions}}.}
}
\description{
Generic plot function
}
