% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppred_examples_df.R
\docType{class}
\name{ppred_examples_df}
\alias{ppred_examples_df}
\title{ppred_examples_df object}
\description{
object of type \code{\link{ppred_examples_df}} is a transformed event log returned by \code{prepare_examples_dt()}.
}
