% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppred_model.R
\docType{class}
\name{ppred_model}
\alias{ppred_model}
\title{ppred_model object}
\description{
object of type \code{\link{ppred_model}} is a list returned by processpredictR::create_model() containing a custom keras functional (transformer) model and some other useful metrics of an event log.
}
