% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppred_predictions.R
\docType{class}
\name{ppred_predictions}
\alias{ppred_predictions}
\title{ppred_predictions object}
\description{
object of type \code{\link{ppred_predictions}} is a data.frame with predicted values returned by predict.ppred_model().
}
