% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_model}
\alias{predict_model}
\title{Predict model}
\usage{
predict_model(transformer_model, tokens_test, predict_type = "y_pred")
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{tokens_test}{A list of test tokens, i.e. token_x, token_y}

\item{predict_type}{A type of output to generate c("y_pred", "metrics")}
}
\description{
(WIP)
}
