% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model_pyversion.R
\name{predict_model_pyversion}
\alias{predict_model_pyversion}
\title{Predict model}
\usage{
predict_model_pyversion(transformer_model, test_data, metrics = FALSE)
}
\arguments{
\item{transformer_model}{A defined transformer model.}

\item{test_data}{A test dataset.}

\item{metrics}{\code{\link{logical}} (default \code{\link{FALSE}}): Returns a vector of predicted values.}
}
\description{
(WIP)
}
