% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ProcessTransformer.R
\name{predict_ProcessTransformer}
\alias{predict_ProcessTransformer}
\title{Generic predict function for ppred_model class.}
\usage{
predict_ProcessTransformer(object, test_data, ...)
}
\arguments{
\item{object}{\code{\link{ppred_model}} (default \code{\link{NULL}}): ProcessTransformer model of class \code{\link{ppred_model}}.}

\item{test_data}{\code{\link{ppred_examples_df}} (default \code{\link{NULL}}): preprocessed test data.}

\item{...}{Additional arguments}
}
\description{
wip
}
