% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_remaining_trace2.R
\name{prep_remaining_trace2}
\alias{prep_remaining_trace2}
\title{Prepares processed event log of class \code{ppred_examples_df} for another variation of remaining trace prediction}
\usage{
prep_remaining_trace2(log)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}
}
\description{
This variation includes as well encoder as decoder block.
}
