% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_examples.R
\name{prepare_examples}
\alias{prepare_examples}
\title{Preprocess an event log.}
\usage{
prepare_examples(
  log,
  task = c("outcome", "next_activity", "next_time", "remaining_time", "remaining_trace"),
  ...
)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},
\code{\link{activitylog}}, etc.).}

\item{task}{\code{\link{character}} (default \code{"outcome"}): A prediction task (\code{"outcome"}, \code{"next_activity"},
\code{"next_time"},\code{"remaining_time"}, \code{"remaining_trace"}).}

\item{...}{Assign outcome labels to each case based on end activity. If not specified, simply defines
outcome as last activity in a trace.}
}
\description{
Prepare an event log for a transformer prediction task.
}
\examples{
# Outcome task prediction with custom outcome labels:
library(eventdataR)
acts <- patients \%>\% edeaR::end_activities("activity")
acts <- unique(acts$activity) \%>\% as.character()
prepare_examples(eventdataR::patients, task = "outcome", outcome_label1 = "Check-out",
outcome_label2 = acts[-1])

# Example next_activity task prediction on traffic_fines dataset:
prepare_examples(traffic_fines, prediction = "next_activity")

# Each prediction task for the patients dataset:
tasks <- c("outcome", "next_activity", "next_time", "remaining_time", "remaining_trace")
purrr::map(tasks, ~prepare_examples(eventdataR::patients, task = .x))

}
