% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_examples_df.R
\name{prepare_examples_dt}
\alias{prepare_examples_dt}
\title{Title}
\usage{
prepare_examples_dt(
  log,
  task = c("outcome", "next_activity", "next_time", "remaining_time", "remaining_trace"),
  features = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
