% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_log.R
\name{preprocess_log}
\alias{preprocess_log}
\title{Prepare event log for the Transformer model fit}
\usage{
preprocess_log(log, ...)
}
\arguments{
\item{log}{An object of class log}
}
\description{
(WIP)
}
