% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppred_examples_tokens.R
\name{print.ppred_examples_tokens}
\alias{print.ppred_examples_tokens}
\title{Generic print function for tokens}
\usage{
\method{print}{ppred_examples_tokens}(x, ...)
}
\arguments{
\item{x}{Object of class token.}

\item{...}{Additional Arguments}
}
\description{
Generic print function for tokens
}
