% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppred_model.R
\name{print.ppred_model}
\alias{print.ppred_model}
\title{Print methods}
\usage{
\method{print}{ppred_model}(x, ...)
}
\arguments{
\item{x}{\code{\link{ppred_model}}: An object of class \code{ppred_model}.}

\item{...}{Additional Arguments.}
}
\value{
prints a Transformer model from a list returned by \code{create_model()}.
}
\description{
Print methods
}
