% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.token.R
\name{print.token}
\alias{print.token}
\title{Generic print function for tokens}
\usage{
\method{print}{token}(x, ...)
}
\arguments{
\item{x}{Object of class token.}

\item{...}{Additional Arguments}
}
\description{
Generic print function for tokens
}
