% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.ppred_model.R, R/fit.ppred_model.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compile}
\alias{fit}
\title{Default compile function for ProcessTransformer model}
\arguments{
\item{optimizer}{Default optimizer for ppred_model}

\item{loss}{Default loss for ppred_model}

\item{metrics}{Default metrics for ppred_model}

\item{object}{Object of class \link{ppred_model}: A transformer model}

\item{train_data}{A training dataset}

\item{batch_size}{A batch size}

\item{num_epochs}{A number of epochs}

\item{verbose}{A verbose}

\item{callbacks}{\code{\link{list}}: A list of callbacks}

\item{shuffle}{\code{\link{logical}} (default \code{\link{TRUE}}): If \code{\link{TRUE}} shuffles the data}

\item{validation_split}{A ratio to split on}

\item{...}{Additional Arguments}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{keras}{\code{\link[keras:reexports]{compile}}, \code{\link[keras:reexports]{fit}}}
}}

