% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test_df.R
\name{split_train_test_df}
\alias{split_train_test_df}
\title{Splits processed dataframe into train- and/or test dataframes}
\usage{
split_train_test_df(processed_df, ratio = 0.7, type_df = NULL)
}
\arguments{
\item{processed_df}{A preprocessed dataframe from preprocess_log function}

\item{ratio}{A train-test split ratio}

\item{type_df}{A type of dataframe to return c("train", "test")}
}
\value{
if type_df is NULL returns a list of train and test dataframes, otherwise either train or test dataframe by specifying type_df
}
\description{
(WIP)
}
