% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_layer.R
\name{stack_layer}
\alias{stack_layer}
\title{Stacks a keras layer on top of existing model}
\usage{
stack_layer(object, ...)
}
\description{
User friendly interface to add a keras layer on top of existing model.
}
