% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_layers.R
\name{stack_layers}
\alias{stack_layers}
\title{Stacks a keras layer on top of existing model}
\usage{
stack_layers(object, ...)
}
\arguments{
\item{object}{a \code{\link{list}} containing a model returned by \code{create_model()}.}

\item{...}{functions for adding layers by using functional keras API. For example, \code{keras::layer_dense(units=32, activation="relu")}.}
}
\value{
a \code{\link{list}} containing an adapted Transformer model.
}
\description{
User friendly interface to add a keras layer on top of existing model.
}
