% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_compile.R
\name{transformer_compile}
\alias{transformer_compile}
\title{Compile and fit transformer model (test)}
\usage{
transformer_compile(transformer_model, learning_rate)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{learning_rate}{A learning rate of a model}
}
\description{
(test)
}
