% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_fit.R
\name{transformer_fit}
\alias{transformer_fit}
\title{Fit model (test)}
\usage{
transformer_fit(
  transformer_model,
  tokens_train,
  maxlen,
  num_epochs,
  batch_size,
  file
)
}
\arguments{
\item{transformer_model}{A defined transformer model}

\item{tokens_train}{A list of train tokens, i.e. token_x, token_y}

\item{maxlen}{An integer number of the maximum case length (longest trace) in an event log}

\item{num_epochs}{A number of epochs}

\item{batch_size}{A batch size}

\item{file}{Name of saved model (weights)}
}
\description{
(test)
}
