% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_model.R
\name{transformer_model}
\alias{transformer_model}
\title{Define transformer model}
\usage{
transformer_model(processed_df)
}
\arguments{
\item{processed_df}{A processed dataframe from create_prefix_df}
}
\value{
A transformer model
}
\description{
(WIP)
}
