% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_model_log.R
\name{transformer_model_log}
\alias{transformer_model_log}
\title{Create transformer model from log}
\usage{
transformer_model_log(log)
}
\arguments{
\item{log}{An object of class log}
}
\value{
a defined transformer model
}
\description{
An alternative to transformer_model() using only log (WIP)
}
