% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_size.R
\name{vocab_size}
\alias{vocab_size}
\title{Calculate the vocabulary size, i.e. the sum of number of activities, outcome labels and padding keys}
\usage{
vocab_size(log, outcome_names)
}
\arguments{
\item{log}{An object of a class log}

\item{outcome_names}{A character vector of outcome labels}
}
\value{
an Integer number of vocabulary size to define the transformer model
}
\description{
(WIP)
}
