% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outcome_labels.R
\name{assign_outcome_labels}
\alias{assign_outcome_labels}
\title{Assign outcome labels}
\usage{
assign_outcome_labels(processed_df, ...)
}
\arguments{
\item{processed_df}{\code{\link{ppred_examples_df}}: A processed dataset returned by \code{prepare_examples()}.}

\item{...}{A named character vector to assign outcome labels to the end activities of outcome}
}
\description{
A shortcut for creating outcome labels based on its column values.
}
