% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outcome_labels.R
\name{assign_outcome_labels}
\alias{assign_outcome_labels}
\title{Assign column outcome with outcome labels based on last activity from the prefixed dataframe}
\usage{
assign_outcome_labels(prefix_df, ...)
}
\arguments{
\item{prefix_df}{A dataframe with prefixes and last activity column from create_prefix_df function}

\item{...}{A named character vector to assign outcome labels to the end activities}
}
\description{
(WIP)
}
