% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix for}
\usage{
confusion_matrix(predictions, ...)
}
\arguments{
\item{predictions}{\code{\link{ppred_predictions}} An event log with predicted values}
}
\description{
Confusion matrix for
}
