% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix for}
\usage{
confusion_matrix(x, as.tibble = F)
}
\arguments{
\item{x}{\code{\link{ppred_predictions}} An event log with predicted values}

\item{as.tibble}{\code{\link{logical}} (default \code{\link{FALSE}}): if \code{\link{TRUE}} returns a tibble}
}
\description{
Confusion matrix for
}
