% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Define transformer model}
\usage{
create_model(df, custom = FALSE, ...)
}
\arguments{
\item{df}{\code{\link{data.frame}}: A processed \code{\link{data.frame}} from prepare_examples}

\item{custom}{\code{\link{logical}} (default \code{\link{FALSE}}): If \code{\link{TRUE}}, returns a custom model.}

\item{...}{you can pass additional arguments to keras::keras_model (ex.: name)}
}
\description{
Defines the model using the keras functional API.
}
