% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Define transformer model}
\usage{
create_model(processed_df, custom_model_py = NULL, name = NULL)
}
\arguments{
\item{processed_df}{A processed \code{\link{data.frame}} from prepare_examples}
}
\value{
A transformer model
}
\description{
(WIP)
}
