% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Define transformer model}
\usage{
create_model(
  processed_df,
  custom_model_py = c("default", "custom"),
  name = NULL
)
}
\arguments{
\item{processed_df}{\code{\link{data.frame}}: A processed \code{\link{data.frame}} from prepare_examples}

\item{custom_model_py}{\code{\link{character}} (default \code{\link{"default"}}): A default or a custom .py model}

\item{name}{\code{\link{character}}: A name for the model}
}
\value{
A transformer model
}
\description{
(WIP)
}
