% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prefix_df.R
\name{create_prefix_df}
\alias{create_prefix_df}
\title{Create a dataframe with prefixes}
\usage{
create_prefix_df(log)
}
\arguments{
\item{log}{An object of class log}
}
\value{
a dataframe
}
\description{
Create a dataframe with prefixes
}
\examples{
library(eventdataR)
create_prefix_df(patients)

}
