% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vocabulary.R
\name{create_vocabulary}
\alias{create_vocabulary}
\title{Create a vocabulary}
\usage{
create_vocabulary(processed_df)
}
\arguments{
\item{processed_df}{A preprocessed object of type \code{\link{log}} or \code{\link{data.frame}} from prepare_examples function.}
}
\value{
A \code{\link{list}} consisting of:
\itemize{
\item {\code{"keys_x"}}: \code{\link{list}} of activity labels
\item {\code{"keys_y"}}: \code{\link{list}} of outcome labels (none for tasks \code{"next_time"} and \code{"remaining_time"})
}
}
\description{
Creates a vocabulary of activities and outcome labels.
}
\examples{
tasks <- c("outcome", "next_activity", "next_time", "remaining_time", "remaining_trace")
purrr::map(tasks, ~prepare_examples(patients, task = .x)) \%>\% purrr::map(create_vocabulary)

}
