% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detokenize.R
\name{detokenize}
\alias{detokenize}
\title{Detokenize predicted values}
\usage{
detokenize(y_pred, y_actual = NULL, vocabulary)
}
\arguments{
\item{y_pred}{\code{\link{array}} An array of predicted values.}

\item{y_actual}{\code{\link{array}} (default \code{\link{NULL}}) An array of actual values.}
}
\description{
Transforms predicted y_values back from tokens to labels
}
