% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detokenize.R
\name{detokenize}
\alias{detokenize}
\title{Detokenize predicted values}
\usage{
detokenize(processed_df, y_pred, y_actual)
}
\arguments{
\item{processed_df}{A preprocessed object of type \code{\link{log}} or \code{\link{data.frame}} from prepare_examples function.}

\item{y_pred}{\code{\link{array}} A vector of predicted values.}

\item{y_actual}{\code{\link{array}} A vector of actual values.}
}
\description{
Transforms predicted y_values back from tokens to labels
}
