% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_attributes.R
\name{map_attributes}
\alias{map_attributes}
\title{Map attributes to a (custom) model}
\usage{
map_attributes(model, processed_df)
}
\arguments{
\item{object}{Object of class \link{ppred_model}}

\item{df}{\code{\link{data.frame}}: A processed \code{\link{data.frame}} from prepare_examples}
}
\description{
Maps attributes and class \code{\link{"ppred_model"}} of a processed \code{\link{data.frame}} to a (custom) model.
}
