% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_case_length.R
\name{max_case_length}
\alias{max_case_length}
\title{Calculate the maximum length of a case / number of activities in the longest trace in an event log}
\usage{
max_case_length(processed_df)
}
\arguments{
\item{processed_df}{A processed dataset of class \code{\link{ppred_examples_df}} returned by \code{prepare_examples()}.}
}
\value{
An \code{integer} number of the maximum case length (longest trace) in an event log.
}
\description{
Calculate the maximum length of a case / number of activities in the longest trace in an event log
}
\examples{
library(processpredictR)
library(eventdataR)

df <- prepare_examples(patients)
max_case_length(df)

}
