% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_outputs.R
\name{num_outputs}
\alias{num_outputs}
\title{Calculate number of outputs (target variables)}
\usage{
num_outputs(processed_df)
}
\arguments{
\item{processed_df}{A processed dataframe}
}
\value{
an integer number of outputs
}
\description{
(WIP)
}
