% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ppred_predictions.R
\name{plot.ppred_predictions}
\alias{plot.ppred_predictions}
\title{Plot Methods}
\usage{
\method{plot}{ppred_predictions}(x, ...)
}
\arguments{
\item{x}{Data to plot. An object of type \code{\link{ppred_predictions}}.}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize metric
}
\concept{visualization}
