% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_log.R
\name{preprocess_log}
\alias{preprocess_log}
\title{Prepare event log for the Transformer model fit}
\usage{
preprocess_log(log, ...)
}
\arguments{
\item{log}{An object of class log}
}
\description{
(WIP)
}
\examples{
acts <- patients \%>\% end_activities("activity") # look on distributions of end activities per case
acts <- unique(acts$activity) \%>\% as.character()
preprocess_log(patients, outcome_label1 = "Check-out", outcome_label2 = acts)

}
