% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_log.R
\name{preprocess_log}
\alias{preprocess_log}
\title{Prepare event log for the Transformer model fit}
\usage{
preprocess_log(log, ...)
}
\arguments{
\item{log}{An object of class log}

\item{...}{A named character vector to assign outcome labels to the end activities}
}
\description{
(WIP)
}
\examples{
acts <- eventdataR::patients \%>\% edeaR::end_activities("activity")
acts <- unique(acts$activity) \%>\% as.character()
preprocess_log(eventdataR::patients, outcome_label1 = "Check-out", outcome_label2 = acts[-1])

}
