% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ppred_model.R
\name{print.ppred_model}
\alias{print.ppred_model}
\title{Print methods}
\usage{
\method{print}{ppred_model}(x, ...)
}
\arguments{
\item{x}{Object of class token.}

\item{...}{Additional Arguments}
}
\description{
Print methods
}
