% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processpredictR.R
\docType{package}
\name{processpredictR}
\alias{processpredictR}
\alias{processpredictR-package}
\title{processpredictR}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
