% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processpredictR.R
\docType{package}
\name{processpredictR}
\alias{processpredictR}
\alias{processpredictR-package}
\title{processpredictR}
\description{
Means to predict process flow, such as process outcome, next activity, next time, remaining time, and remaining trace. Off-the-shelf predictive models based on the concept of Transformers are provided, as well as multiple ways to customize the models. This package is partly based on work described in Zaharah A. Bukhsh, Aaqib Saeed, & Remco M. Dijkman. (2021). “ProcessTransformer: Predictive Business Process Monitoring with Transformer Network”.\href{https://arxiv.org/abs/2104.00721}{arXiv:2104.00721}
}
\author{
\strong{Maintainer}: Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be}

Authors:
\itemize{
  \item Ivan Esin \email{ivan.esin@student.uhasselt}
}

Other contributors:
\itemize{
  \item Hasselt University [copyright holder]
}

}
