% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.ppred_model.R,
%   R/evaluate.ppred_model.R, R/fit.ppred_model.R, R/predict.ppred_model.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compile}
\alias{evaluate}
\alias{fit}
\alias{predict}
\title{Default compile function for ProcessTransformer model}
\arguments{
\item{optimizer}{Default optimizer for ppred_model}

\item{loss}{Default loss for ppred_model}

\item{metrics}{Default metrics for ppred_model}

\item{train_data}{A training dataset}

\item{batch_size}{A batch size}

\item{num_epochs}{A number of epochs}

\item{verbose}{A verbose}

\item{callbacks}{\code{\link{list}}: A list of callbacks. \code{keras} default is \code{NULL}, but can be adjusted (ex. keras::callback_csv_logger(filename = paste("log_", object$task)), #or NULL
keras::callback_tensorboard())}

\item{shuffle}{\code{\link{logical}} (default \code{\link{TRUE}}): If \code{\link{TRUE}} shuffles the data}

\item{validation_split}{A ratio to split on}

\item{object}{\code{\link{ppred_model}} (default \code{\link{NULL}}): ProcessTransformer model of class \code{\link{ppred_model}}.}

\item{test_data}{\code{\link{ppred_examples_df}} (default \code{\link{NULL}}): preprocessed test data.}

\item{append}{\code{\link{logical}} (default \code{\link{FALSE}}): if \code{\link{TRUE}}, returns a passed \code{\link{data.frame}} with predicted values.}

\item{...}{Additional arguments}
}
\seealso{
See \code{\link[keras:reexports]{keras::fit()}} for documentation of parameters
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{keras}{\code{\link[keras:reexports]{compile}}, \code{\link[keras:reexports]{evaluate}}, \code{\link[keras:reexports]{fit}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

