% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.ppred_model.R, R/fit.ppred_model.R,
%   R/predict.ppred_model.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{evaluate}
\alias{fit}
\alias{predict}
\title{Generic evaluate function for ppred_model class.}
\arguments{
\item{train_data}{A training dataset}

\item{batch_size}{A batch size}

\item{num_epochs}{A number of epochs}

\item{verbose}{A verbose}

\item{callbacks}{\code{\link{list}}: A list of callbacks}

\item{shuffle}{\code{\link{logical}} (default \code{\link{TRUE}}): If \code{\link{TRUE}} shuffles the data}

\item{validation_split}{A ratio to split on}

\item{object}{\code{\link{ppred_model}} (default \code{\link{NULL}}): ProcessTransformer model of class \code{\link{ppred_model}}.}

\item{test_data}{\code{\link{ppred_examples_df}} (default \code{\link{NULL}}): preprocessed test data.}

\item{...}{Additional arguments}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{keras}{\code{\link[keras:reexports]{evaluate}}, \code{\link[keras:reexports]{fit}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

