% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.ppred_model.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compile}
\title{Default compile function for ProcessTransformer model}
\arguments{
\item{object}{Object of class \link{ppred_model}}

\item{...}{Additional Arguments}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{keras}{\code{\link[keras:reexports]{compile}}}
}}

