% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{Splits processed dataframe into train- and/or test dataframes}
\usage{
split_train_test(processed_df, split = 0.7, trace_length_bins = 5)
}
\arguments{
\item{processed_df}{A preprocessed dataframe from preprocess_log function}

\item{split}{A train-test split ratio}

\item{trace_length_bins}{Number of trace length bins to use for stratification}
}
\value{
a list of train and test dataframes
}
\description{
(WIP)
}
