% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_layers.R
\name{stack_layers}
\alias{stack_layers}
\title{Stacks a keras layer on top of existing model}
\usage{
stack_layers(object, ...)
}
\arguments{
\item{object}{a list containing a model returned by create_model().}

\item{...}{functions for adding layers by using functional keras API. For example, keras::layer_dense(units=32, activation="relu")}
}
\description{
User friendly interface to add a keras layer on top of existing model.
}
