% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_layers.R
\name{stack_layers}
\alias{stack_layers}
\title{Stacks a keras layer on top of existing model}
\usage{
stack_layers(object, ...)
}
\description{
User friendly interface to add a keras layer on top of existing model.
}
