% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize prefix (feature) and outcome (target) of a processed dataframe to fit the transformer model}
\usage{
tokenize(processed_df)
}
\arguments{
\item{processed_df}{A processed dataframe (= processed event log)}

\item{activities}{A character vector of unique activity labels from an event log}
}
\value{
A list of tokens token_x and token_y
}
\description{
(WIP)
}
