% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer_model.R
\name{transformer_model}
\alias{transformer_model}
\title{Define transformer model}
\usage{
transformer_model(max_case_length, vocab_size, num_output)
}
\arguments{
\item{max_case_length}{An integer number which is the maximum trace length in an event log}

\item{vocab_size}{An integer number which is the number of unique activities, output labels and padding keys ("PAD", "UNK")}

\item{num_output}{An integer number of output labels}
}
\value{
A transformer model
}
\description{
(WIP)
}
