% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_size.R
\name{vocab_size}
\alias{vocab_size}
\title{Calculate the vocabulary size, i.e. the sum of number of activities, outcome labels and padding keys}
\usage{
vocab_size(processed_df)
}
\arguments{
\item{processed_df}{A processed dataset of class \code{\link{ppred_examples_df}} from \code{prepare_examples()}.}
}
\value{
an \code{integer} number of vocabulary size to define the Transformer model.
}
\description{
Calculate the vocabulary size, i.e. the sum of number of activities, outcome labels and padding keys
}
\examples{
library(processpredictR)
library(eventdataR)
df <- prepare_examples(patients)
vocab_size(df)

}
